Change to install-dotfiles.sh to make it more amenable to running on macOS

+++ bin/install-dotfiles.sh
@@ -1,7 +1,10 @@
-source /etc/os-release
-if [ ${ID} = "ubuntu" -o ${ID} = "debian" ]
+if [[ -e /etc/os-release ]]
 then
-    sudo apt-get --yes install curl
+    source /etc/os-release
+    if [ ${ID} = "ubuntu" -o ${ID} = "debian" ]
+    then
+        sudo apt-get --yes install curl
+    fi
 fi
 export github_user=jafingerhut
 bash -c "$(curl -fsSL https://raw.github.com/$github_user/dotfiles/master/bin/dotfiles)" && source ~/.bashrc
